Adds secondary button widget
Implements a new secondary button widget with outlined style.

This commit introduces `AppSecondaryButton`, a new button widget that provides an outlined button style. It supports both dark (white border) and light (gray border) themes based on the `isDark` property. The button also includes customizable text and optional child widgets.

Updates the primary button to use a sized box and removes the custom text widget. It also updates the button style to use a rounded rectangle with a radius of 25 and sets the text color to white.