Implements cooperative selection sidebar
Adds a cooperative selection sidebar accessible via a drawer.

The sidebar allows users to switch between different cooperatives.
It includes a toggle button on the top left corner of the screen
to open the drawer. Styles the header to accommodate the new toggle button and
make the cooperative name and user info more compact and readable.